Unified Makefile target for Atari target and macOS host

+++ Makefile
@@ -1,35 +1,53 @@
 include ../toybox/common.mk
+GAME=.
+GAMEINC=$(GAME)/include
 
-CG_OBJECTS=build/main.o
-CG_OBJECTS+=build/resources.o build/level.o build/button.o build/scroller.o
-CG_OBJECTS+=build/game.o build/game_menu.o build/game_scores.o build/game_credits.o build/game_help.o build/game_level.o build/game_level_edit.o
+CFLAGS+=-I $(GAMEINC)
+
+# Source files for cgrid
+SOURCES=$(wildcard src/*.cpp)
+OBJECTS=$(patsubst src/%.cpp,build/%.o,$(SOURCES))
 
 all: cgrid
 
 toybox:
-	$(MAKE) -C $(TOYBOX) libtoybox.a
+	$(MAKE) -C $(TOYBOX) libtoybox.a $(if $(HOST),HOST=$(HOST))
 
-cgrid: toybox $(CG_OBJECTS)
-	$(CC) $(LIBCMINILIB)/startup.o $(CG_OBJECTS) $(LDFLAGS) -o build/cgrid.tos
+cgrid: toybox $(OBJECTS)
+ifeq ($(HOST),sdl2)
+	$(CC) $(OBJECTS) $(LDFLAGS) -o build/cgrid
+else
+	$(CC) $(LIBCMINILIB)/startup.o $(OBJECTS) $(LDFLAGS) -o build/cgrid.tos
+endif
 
 build/%.o: src/%.cpp
 	$(CC) $(CFLAGS) $< -o $@
 
 install: cgrid
 	mkdir -p install
+ifeq ($(HOST),sdl2)
+	cp build/cgrid install/cgrid
+else
 	mkdir -p install/auto
 	cp build/cgrid.tos install/cgrid.tos
 	cp build/cgrid.tos install/auto/cgrid.prg
-	cp -r data install/data
-	cp other/readme.txt install/readme.txt
 	cp -r other/bonus install/bonus
+	cp other/readme.txt install/readme.txt
+endif
+	cp -r data install/data
 
 cracked: install
+ifeq ($(HOST),sdl2)
+	@echo "cracked target not supported for SDL2 build"
+else
 	rm install/auto/cgrid.prg
 	cp other/d-bug install/d-bug
 	cp other/loader.prg install/auto/loader.prg
+endif
 
 clean:
 	$(MAKE) -C $(TOYBOX) clean
 	rm -rf build
 	rm -rf install
+
+.PHONY: all toybox install cracked clean
